Replace is_a? type checks with === checks
This allows us to use any object that responds to ===
for checking value conformance. This is useful in combination
with, say, the "mal" gem or with Regexes after applying
typecasts. This also allows us to use a library that defines
union types etc. for doing more complex match checks as
opposed to a single inheritance check.