Preserve undocumented route params
Previously, if you would define a Sinatra route
using "/things/:id" without adding a `route_param`
using Apiculture, this param would be stripped both
from the route arguments AND from the Sinatra params
hash. Since the route params will always be present
anyway, we can just let them pass. It is not cool
to not document them, but not documenting them must
not be punished with a silent removal.